modify navTabs and transfer timer
